% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_cleaner.R
\name{address_cleaner}
\alias{address_cleaner}
\title{Performs character transformations on a vector of
addresses in order to build "web-safe" URLs for the Google API.}
\usage{
address_cleaner(address, verbose = TRUE)
}
\arguments{
\item{address}{A raw 1xN vector of UTF-8 encoded addresses.
Note: these addresses should be in raw form, \emph{not} URL encoded (e.g., of the form: 123 Main Street, Somewhere, NY 12345 USA)(country is optional but recommended).}

\item{verbose}{Displays additional progress output}
}
\description{
This function scrubs a vector of addresses (e.g., a vector of the form: address, city, state, postal code, country)
of character values that may inhibit sucessful geocoding with the \href{https://developers.google.com/maps/documentation/geocoding/start}{Google maps API}.
Specifically, this function:
\itemize{
  \item {Replaces non-breaking spaces with " "}
  \item {Removes ASCII control characters (001-031 and 177)}
  \item {Trims runs of spaces and spaces which begin/end a string}
  \item {Converts special addressing charecters, such as ordinals}
  \item {Strips latin1 characters}
  \item {Removes leading, trailing, and repeated commas}
  \item {Remove various permutations of the "c/o" flag}
  }
}
\details{
Note: We recommend reviewing the output of this function against the original vector to ensure the
  transformations it performs resolve the formatting issues (if any) found in your source data.
}

