% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_cleaner.R
\name{address_cleaner}
\alias{address_cleaner}
\title{Performs common character transformations to a data frame of
addresses in order to build "web-safe" URLs for the Google API.}
\usage{
address_cleaner(address, verbose = TRUE)
}
\arguments{
\item{address}{A data frame of addresses fields, generally of the form: address, city, state, postal code, country.}

\item{verbose}{Displays additional progress output}
}
\description{
This function scrubs a data frame of addresses (e.g., one of the form: address, city, state, postal code, country)
of character values that may inhibit sucessful geocoding with the \href{https://developers.google.com/maps/documentation/geocoding/start}{Google maps API}.
Specifically, this function:
\itemize{
  \item {Replaces non-breaking spaces with " "}
  \item {Removes ASCII control characters (001-031 and 177)}
  \item {Trims runs of spaces and spaces which begin/end a string}
  \item {Converts special addressing charecters, such as ordinals}
  \item {Strips latin1 characters}
  \item {Removes leading, trailing, and repeated commas}
  \item {Remove various permutations of the "c/o" flag}
  }
}
\details{
Note: We recommend closely reviewing the output of this function against the original to ensure that these
  transformations comport with the issues (if any) found in your address list.
}

