% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_url.R
\name{geocode_url}
\alias{geocode_url}
\title{Geocode an address vector using the (paid) Google for Work API.}
\usage{
geocode_url(address, verbose = FALSE, add_date = TRUE, fuzzy = FALSE)
}
\arguments{
\item{address}{A "url-safe" vector of fully formed URLs that are encoded with your private Google key (see: \code{\link{google_encode64}}).}

\item{verbose}{Display additional output in the returns from Google?}

\item{add_date}{Adds a column with today's date to the returned data.frame.}

\item{fuzzy}{Uses \code{\link{runif}} to add a random number of days between 1 and 30 to the current date when add_date is TRUE. This can be useful to avoid sending large batches if your scripts recertify/retry geocoding after a fixed number of days.}
}
\description{
This function attempts to geocode a vector of physical addresses using Google's
(paid) \href{https://www.google.com/work/}{Google for Work API}. A version of this
function is discussed in # A version of this function is discussed
\href{http://stackoverflow.com/questions/3257441/geocoding-in-r-with-google-maps}{in this thread}.
This modified version of the function retains additional parameters in the JSON-formatted return:
modified version adds to parameters from the JSON return:
\itemize{
  \item \emph{location_type}
  \item \emph{status}
  }
}

