% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_encode64.R
\name{google_encode64}
\alias{google_encode64}
\title{Build fully formed URLs for geocoding using one's Google for Work private API key.}
\usage{
google_encode64(address, privkey, clientid, sigfile = FALSE,
  sigfile_loc = NULL, verbose = FALSE)
}
\arguments{
\item{address}{A 1xN locations vector with UTF-8 encoded ASCII-characters (see: ).}

\item{privkey}{Your Google for Work API key}

\item{clientid}{This parameter is generall of the form 'gme-[company]'}

\item{sigfile}{Save out a complete file of locations and URL signatures (this can be useful for debugging malformed URLs)?}

\item{sigfile_loc}{A fully formed folder/file location to save the function's output. If NULL and sigfile=TRUE, file is saved as "signature_file_[date-time].Rds" in the current working directory.}

\item{verbose}{Displays additional progress output}
}
\description{
This function creates a fully formed vector of URLs encoded with your
\href{https://www.google.com/work/}{Google for Work} private API key.
The general pattern of this function follows the \href{https://developers.google.com/maps/documentation/business/image/auth?hl=en}{Google Developer documention}
for generating a digital signature. Prior to applying this function you may need to strip your address vector of non-URL-safe string values (see: ).
}

